Fixed Deletion Issues

+++ Intex2/Controllers/BurialCrudController.cs
@@ -95,8 +95,8 @@ namespace Intex2.Controllers
 
             if (ModelState.IsValid)
             {
-                //_context.Add(burials);
-                //await _context.SaveChangesAsync();
+                _context.Add(burials);
+                await _context.SaveChangesAsync();
                 string referenceId = burials.BurialId;
                 return View("Details", burials);
             }
@@ -197,8 +197,14 @@ namespace Intex2.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(string id)
         {
+            string newid = id.Replace("%2F", "/");
 
-            var burials = await _context.Burials.FindAsync(id);
+            if (newid == null)
+            {
+                return NotFound();
+            }
+
+            var burials = await _context.Burials.FindAsync(newid);
             _context.Burials.Remove(burials);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
